requirements.txt
psutil==6.1.0
python_snap7==2.0.2
scapy==2.6.0
